Updated TeamAdvanced, TeamFav, and TeamController, both team API's function properly

+++ NFLWeatherAppAPI/Entities/TeamAdvancedStats.cs
@@ -17,10 +17,10 @@ namespace NFLWeatherAppAPI.Entities
 
         public string Division { get; set; }
 
-        public int PointsPerGame { get; set; }
+        public double PointsPerGame { get; set; }
         [Required]
 
-        public int YardsPerGame { get; set; }
+        public double YardsPerGame { get; set; }
         
     }
 }
